V2.0.0
It is now possible to leave a BMS in Master mode.

This involves setting all 4 small switches to the **down** or **0000** position.
Parameters can no longer be modified in this mode !

In this configuration, the BMS will query the other BMSs if any are connected to the RS485 bus.

The BMS will then broadcast the information it has retrieved cyclically every 5 seconds. This information will be read by this add-on